+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 * Support for minimum and maximum hours for timepicker.
 * New `setDate()` method for single datepicker. The new method combines `setStartDate()` and `setEndDate()` methods in one call.
 * New `setMinDate()` method.
+* Calendar titles.
 
 ### Changed